+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Error Message - Invoice cancellation allowed against closed budget. Refs UINV-513.
 * Not able to filter with tags that are entirely made of special characters. Refs UINV-510.
 * Refactor CSS away from postcss-color-function. Refs UINV-516.
+* Display all external account numbers associated with an invoice in the export report. Refs UINV-515.
 
 ## [5.0.1](https://github.com/folio-org/ui-invoice/tree/v5.0.1) (2023-10-31)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v5.0.0...v5.0.1)